requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlmodel==0.0.14
python-dotenv==1.0.1
qdrant-client==1.7.3
requests==2.31.0
tenacity==8.2.3 